+++ verifyCredentials.ts
@@ -1,0 +1,15 @@
+import Sftp from './src/Sftp';
+
+export = async function verifyCredentials(cfg: any) {
+  const sftp = new Sftp(this.logger, cfg);
+  try {
+    await sftp.connect();
+    await sftp.end();
+    this.logger.info('SFTP credentials successfully verified');
+    return { verified: true };
+  } catch (error) {
+    this.logger.error(error.message);
+    this.logger.error('SFTP credentials verification failed');
+    return { verified: false };
+  }
+}